requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.2
Django==6.0.2
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-js-asset==3.1.2
django-widget-tweaks==1.5.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.0
psycopg2==2.9.11
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
